History files for all nations missing them
Turns out having a blank history files throws errors in the log. Also,
countries without history files can't be released as vassals, and crash
the game when clicked on. Long story short, every country should have a
history file, even if it doesn't exist at the start of the game.