Merged Mac Monica's fork
"I've been removing the converter files as we no longer need them and correcting the names of provinces or adding new provinces as needed. Some of the regions were named after colonial boundaries from the 15th century for example and I've added some more detail so the bretheran can have more of a presence. I've done the region of the Caribbean and central america up to the Yucatan."